+++ src/expressions/boolean_not/parse.ts
@@ -30,6 +30,7 @@ import {
   TOKEN_GT,
   TOKEN_IDENT,
   TOKEN_INTEGER,
+  TOKEN_LBRACKET,
   TOKEN_LE,
   TOKEN_LG,
   TOKEN_LPAREN,
@@ -81,6 +82,7 @@ export const TOKEN_MAP = new Map<string, parseFunc>([
   [TOKEN_INTEGER, parseIntegerLiteral],
   [TOKEN_FLOAT, parseFloatLiteral],
   [TOKEN_IDENT, parseIdentifier],
+  [TOKEN_LBRACKET, parseIdentifier],
   [TOKEN_RANGE_LPAREN, makeParseRange(parseObject, TOKEN_RANGE_LPAREN)],
 ]);